requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
